Integrating set-udf with ordinary assignment
Update to db reflects the state of the domain object. It's also possible to update udf that is not represented as an instance variable in the corresponding domain object. The function set_udf updates instance variable if it exists, otherwise, updates the api resource. When committing, all changes in instance variables are transferred to the api resources.